+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.32] - 2026-06-05
+
 ### Added
 
 - `dew assets pull --force` re-downloads kernel + initramfs even when